requirements.txt
feedparser
pyyaml
requests
python-dateutil
google-genai